requirements.txt
flask
feedparser
requests
readability-lxml
ebooklib
beautifulsoup4
Pillow
html5-parser
lxml
python-dotenv
